(chore): fix minor bugs, fix circular dependency, update yarn
