+++ packages/sdk/package.json
@@ -84,6 +84,7 @@
     "numeral": "^2.0.6",
     "rollup-plugin-multi-input": "^1.3.1",
     "web3": "^1.7.1",
+    "web3-eth": "^1.7.1",
     "web3-eth-contract": "^1.7.1",
     "web3-utils": "^1.4.0"
   }
