+++ packages/sdk/src/core/buy.ts
@@ -15,11 +15,9 @@ import { Trade } from '@uniswap/v2-sdk'
 import { getToken } from 'methods/tokenLists'
 import { decimalPercentToPercent, decimalToJSBI } from 'utils/converter'
 import { CDAI, FUSE } from 'constants/tokens'
-import { cDaiPrice } from 'methods/cDaiPrice'
 import { v2TradeExactIn } from 'methods/v2TradeExactIn'
-import { goodMarketMakerContract } from 'contracts/GoodMarketMakerContract'
 import { getAccount, getChainId } from 'utils/web3'
-import { UnexpectedToken, UnsupportedChainId, UnsupportedToken } from 'utils/errors'
+import { UnsupportedChainId, UnsupportedToken } from 'utils/errors'
 import { debug, debugGroup, debugGroupEnd } from 'utils/debug'
 import { ZERO_PERCENT } from 'constants/misc'
 import { exchangeHelperContract } from 'contracts/ExchangeHelperContract'
@@ -27,38 +25,9 @@ import { prepareValues } from 'functions/prepareValues'
 import { computeRealizedLPFeePercent } from 'utils/prices'
 import { SupportedChainId } from 'constants/chains'
 import { v2TradeExactOut } from 'methods/v2TradeExactOut'
-// eslint-disable-next-line import/no-cycle
-import { cDaiToDai, G$ToCDai } from './sell'
 import * as fuse from 'contracts/FuseUniswapContract'
 import { g$ReservePrice } from 'methods/g$price'
-
-export type BuyInfo = {
-    inputAmount: CurrencyAmount<Currency>
-    outputAmount: CurrencyAmount<Currency>
-    minimumOutputAmount: CurrencyAmount<Currency>
-
-    DAIAmount: CurrencyAmount<Currency> | null
-    cDAIAmount: CurrencyAmount<Currency> | null
-    GDXAmount: CurrencyAmount<Currency> | Fraction
-
-    priceImpact: Fraction
-    slippageTolerance: Percent
-
-    liquidityFee: CurrencyAmount<Currency>
-    liquidityToken: Currency
-
-    route: Token[]
-    trade: Trade<Currency, Currency, TradeType> | null
-}
-
-type DAIResult = {
-    amount: CurrencyAmount<Currency>
-    minAmount: CurrencyAmount<Currency>
-    route: Token[]
-    trade: Trade<Currency, Currency, TradeType>
-}
-
-type G$Result = Omit<DAIResult, 'route' | 'trade'>
+import { cDaiToDai, G$ToCDai, daiToCDai, cDaiToG$, SwapInfo as BuyInfo, XResult as DAIResult} from './swap'
 
 /**
  * Tries to convert token X into DAI. If it impossible - returns null.
@@ -191,78 +160,6 @@ export async function xToG$ExactOut(
     debugGroupEnd(`${currency.symbol} to G$`)
 
     return trade.inputAmount
-}
-
-/**
- * Tries to convert token DAI into cDAI.
- * @param {Web3} web3 Web3 instance.
- * @param {CurrencyAmount<Currency>} currency DAI token amount.
- * @returns {CurrencyAmount<Currency>}
- * @throws {UnexpectedToken} If currency not DAI.
- */
-export async function daiToCDai(web3: Web3, currency: CurrencyAmount<Currency>): Promise<CurrencyAmount<Currency>> {
-    if (currency.currency.symbol !== 'DAI') {
-        throw new UnexpectedToken(currency.currency.symbol)
-    }
-
-    const chainId = await getChainId(web3)
-
-    debugGroup(`DAI to cDAI`)
-
-    const cDaiPriceRatio = await cDaiPrice(web3, chainId)
-    debug('cDAI ratio', cDaiPriceRatio.toSignificant(6))
-
-    // DAI is 18 decimal number, cDAI is 8 decimal number, need to reduce 10 - 8 = 10 decimals from final value
-    const _cDaiOutput = currency.divide(cDaiPriceRatio).divide(1e10)
-    const amount = CurrencyAmount.fromFractionalAmount(CDAI[chainId], _cDaiOutput.numerator, _cDaiOutput.denominator)
-
-    debug('cDAI', amount.toSignificant(6))
-    debugGroupEnd(`DAI to cDAI`)
-
-    return amount
-}
-
-let index = 0
-
-/**
- * Tries to convert token cDAI into G$.
- * @param {Web3} web3 Web3 instance.
- * @param {CurrencyAmount<Currency>} currency CDAI token amount.
- * @param {Percent} slippageTolerance Slippage tolerance.
- * @returns {G$Result}
- * @throws {UnexpectedToken} If currency not cDAI.
- */
-export async function cDaiToG$(
-    web3: Web3,
-    currency: CurrencyAmount<Currency>,
-    slippageTolerance: Percent
-): Promise<G$Result> {
-    if (currency.currency.symbol !== 'cDAI') {
-        throw new UnexpectedToken(currency.currency.symbol)
-    }
-
-    const chainId = await getChainId(web3)
-    const goodMarketMaker = await goodMarketMakerContract(web3)
-    const G$ = (await getToken(chainId, 'G$')) as Token
-
-    const _index = ++index
-    debugGroup(`cDAI to G$ - ${_index}`)
-
-    const bigNumber = currency.multiply(currency.decimalScale).toFixed(0)
-    const _priceMinimumOutputAmount = (await goodMarketMaker.methods
-        .buyReturn(CDAI[chainId].address, bigNumber)
-        .call()) as BigNumber
-
-    const amount = CurrencyAmount.fromRawAmount(G$, _priceMinimumOutputAmount.toString())
-    debug('G$', amount.toSignificant(6))
-
-    const minAmount = amount.subtract(amount.multiply(slippageTolerance))
-    debug('G$ min', minAmount.toSignificant(6))
-
-    debugGroupEnd(`cDAI to G$ - ${_index}`)
-    index--
-
-    return { amount, minAmount }
 }
 
 /**
@@ -501,64 +398,6 @@ export async function getBuyMetaReverse(
 
     return getBuyMeta(web3, fromSymbol, result.toExact(), slippageTolerance)
 }
-
-// /**
-//  * Pick necessary date from meta swap.
-//  * @param {BuyInfo} meta Result of the method getMeta() execution.
-//  * @returns {input: string, minReturn: string, minDai: string}
-//  */
-// function prepareValues(meta: BuyInfo): { input: string; minReturn: string; minDai: string } {
-//     if (!meta.route.length) {
-//         throw new InsufficientLiquidity()
-//     }
-
-//     const input = meta.inputAmount.multiply(meta.inputAmount.decimalScale).toFixed(0)
-//     const minReturn = meta.minimumOutputAmount.multiply(meta.minimumOutputAmount.decimalScale).toFixed(0)
-//     const minDai = meta.DAIAmount ? meta.DAIAmount.multiply(meta.DAIAmount.decimalScale).toFixed(0) : '0'
-
-//     debug({
-//         input: meta.inputAmount.toSignificant(6),
-//         minReturn: meta.minimumOutputAmount.toSignificant(6),
-//         minDai: meta.DAIAmount ? meta.DAIAmount.toSignificant(6) : '0'
-//     })
-
-//     return { input, minReturn, minDai }
-// }
-
-
-
-//@deprecated: replaced by functions/approve
-/**
- * Approve token usage.
- * @param {Web3} web3 Web3 instance.
- * @param {BuyInfo} meta Result of the method getMeta() execution.
- */
-// export async function approveBuy(web3: Web3, meta: BuyInfo): Promise<void> {
-//     const chainId = await getChainId(web3)
-
-//     if (meta.trade && meta.trade.inputAmount.currency.isNative) {
-//         return
-//     } else if (chainId === SupportedChainId.FUSE) {
-//         await fuse.approveBuy(web3, meta.trade!)
-//     } else {
-//         const account = await getAccount(web3)
-//         const { input } = prepareValues(meta)
-//         const bigInput = BigNumber.from(input)
-
-//         const erc20 = ERC20Contract(web3, meta.route[0].address)
-
-//         const allowance = await erc20.methods
-//             .allowance(account, G$ContractAddresses(chainId, 'ExchangeHelper'))
-//             .call()
-//             .then((_: string) => BigNumber.from(_))
-
-//         if (bigInput.lte(allowance)) return
-
-//         await erc20.methods
-//             .approve(G$ContractAddresses(chainId, 'ExchangeHelper'), MaxUint256.toString())
-//             .send({ from: account })
-//     }
-// }
 
 /**
  * Swap tokens.
