+++ packages/sdk/src/core/index.ts
@@ -3,5 +3,6 @@ export * from "./buy";
 export * from "./rest";
 export * from "./sell";
 export * from "./staking";
+export * from "./swap";
 export * from "./tokens";
 export * from "./ubi";
