+++ packages/sdk/src/core/sell.ts
@@ -15,14 +15,9 @@ import { getToken } from 'methods/tokenLists'
 import { calculateExitContribution } from 'methods/calculateExitContribution'
 import { decimalPercentToPercent, decimalToJSBI, g$FromDecimal } from 'utils/converter'
 import { CDAI, FUSE } from 'constants/tokens'
-import { cDaiPrice } from 'methods/cDaiPrice'
 import { v2TradeExactIn } from 'methods/v2TradeExactIn'
-// eslint-disable-next-line import/no-cycle
-import { BuyInfo, cDaiToG$, daiToCDai } from './buy'
-import { goodMarketMakerContract } from 'contracts/GoodMarketMakerContract'
 import { getAccount, getChainId } from 'utils/web3'
 import { debug, debugGroup, debugGroupEnd } from 'utils/debug'
-import { UnexpectedToken } from 'utils/errors'
 import { computeRealizedLPFeePercent } from 'utils/prices'
 import { tokenBalance } from 'utils/tokenBalance'
 import { exchangeHelperContract } from 'contracts/ExchangeHelperContract'
@@ -33,17 +28,9 @@ import { TransactionDetails } from 'constants/transactions'
 import * as fuse from 'contracts/FuseUniswapContract'
 import { g$ReservePrice } from 'methods/g$price'
 import { prepareValues } from 'functions/prepareValues'
+import { cDaiToDai, G$ToCDai, XResult, SwapInfo, cDaiToG$, daiToCDai } from './swap'
 
-export type SellInfo = { contribution: Fraction } & BuyInfo
-
-type XResult = {
-    amount: CurrencyAmount<Currency>
-    minAmount: CurrencyAmount<Currency>
-    route: Token[]
-    trade: Trade<Currency, Currency, TradeType>
-}
-
-type CDAIResult = Omit<XResult, 'route' | 'trade'>
+export type SellInfo = { contribution: Fraction } & SwapInfo
 
 /**
  * Tries to convert token DAI into X. If it impossible - returns null.
@@ -177,72 +164,6 @@ export async function G$ToXExactOut(
     debugGroupEnd(`G$ to ${to.currency.symbol}`)
 
     return trade.inputAmount
-}
-
-/**
- * Tries to convert token cDAI into DAI.
- * @param {Web3} web3 Web3 instance.
- * @param {CurrencyAmount<Currency>} currency CDAI token amount.
- * @returns {CurrencyAmount<Currency>}
- * @throws {UnexpectedToken} If currency not DAI.
- */
-export async function cDaiToDai(web3: Web3, currency: CurrencyAmount<Currency>): Promise<CurrencyAmount<Currency>> {
-    if (currency.currency.symbol !== 'cDAI') {
-        throw new UnexpectedToken(currency.currency.symbol)
-    }
-
-    const chainId = await getChainId(web3)
-    const DAI = (await getToken(chainId, 'DAI')) as Token
-
-    debugGroup(`cDAI to DAI`)
-
-    const cDaiPriceRatio = await cDaiPrice(web3, chainId)
-    debug('cDAI ratio', cDaiPriceRatio.toSignificant(6))
-
-    // DAI is 18 decimal number, cDAI is 8 decimal number, need to add 10 - 8 = 10 decimals from final value
-    const _daiAmount = currency.multiply(cDaiPriceRatio).multiply(1e10)
-    const amount = CurrencyAmount.fromFractionalAmount(DAI, _daiAmount.numerator, _daiAmount.denominator)
-
-    debug('cDAI', amount.toSignificant(6))
-    debugGroupEnd(`cDAI to DAI`)
-
-    return amount
-}
-
-/**
- * Tries to convert token G$ into cDAI.
- * @param {Web3} web3 Web3 instance.
- * @param {CurrencyAmount<Currency>} currency G$ token amount.
- * @param {Percent} slippageTolerance Slippage tolerance.
- * @returns {CDAIResult}
- * @throws {UnexpectedToken} If currency not cDAI.
- */
-export async function G$ToCDai(
-    web3: Web3,
-    currency: CurrencyAmount<Currency>,
-    slippageTolerance: Percent
-): Promise<CDAIResult> {
-    if (currency.currency.symbol !== 'G$') {
-        throw new UnexpectedToken(currency.currency.symbol)
-    }
-
-    const chainId = await getChainId(web3)
-    const goodMarketMaker = await goodMarketMakerContract(web3)
-
-    debugGroup(`G$ to cDAI`)
-
-    const bigNumber = currency.multiply(currency.decimalScale).toFixed(0)
-    const _cDaiOutput = await goodMarketMaker.methods.sellReturn(CDAI[chainId].address, bigNumber).call()
-
-    const amount = CurrencyAmount.fromRawAmount(CDAI[chainId], _cDaiOutput.toString())
-    debug('cDAI', amount.toSignificant(6))
-
-    const minAmount = amount.subtract(amount.multiply(slippageTolerance))
-    debug('cDAI min', minAmount.toSignificant(6))
-
-    debugGroupEnd(`G$ to cDAI`)
-
-    return { amount, minAmount }
 }
 
 /**
@@ -475,71 +396,15 @@ export async function getSellMetaReverse(
     return getSellMeta(web3, toSymbol, result.toExact(), slippageTolerance)
 }
 
-//@deprecated: replaced by functions/prepareValues
-// /**
-//  * Pick necessary date from meta swap.
-//  * @param {BuyInfo} meta Result of the method getMeta() execution.
-//  * @returns {input: string, minReturn: string, minCDai: string}
-//  */
-// function prepareValues(meta: BuyInfo): { input: string; minReturn: string; minCDai: string } {
-//     if (!meta.route.length) {
-//         throw new InsufficientLiquidity()
-//     }
-
-//     const input = meta.inputAmount.multiply(meta.inputAmount.decimalScale).toFixed(0)
-//     const minReturn = meta.minimumOutputAmount.multiply(meta.minimumOutputAmount.decimalScale).toFixed(0)
-//     const minCDai = meta.cDAIAmount ? meta.cDAIAmount.multiply(meta.cDAIAmount.decimalScale).toFixed(0) : '0'
-
-//     debug({
-//         input: meta.inputAmount.toSignificant(6),
-//         minReturn: meta.minimumOutputAmount.toSignificant(6),
-//         minCDai: meta.cDAIAmount ? meta.cDAIAmount.toSignificant(6) : '0'
-//     })
-
-//     return { input, minReturn, minCDai }
-// }
-
-//@deprecated: replaced by functions/approve
-// /**
-//  * Approve token usage.
-//  * @param {Web3} web3 Web3 instance.
-//  * @param {BuyInfo} meta Result of the method getMeta() execution.
-//  */
-// export async function approve(web3: Web3, meta: BuyInfo): Promise<void> {
-//     const chainId = await getChainId(web3)
-
-//     if (chainId === SupportedChainId.FUSE) {
-//         await fuse.approveSell(web3, meta.trade!)
-//     } else {
-//         const account = await getAccount(web3)
-//         const { input } = prepareValues(meta)
-//         const bigInput = BigNumber.from(input)
-
-//         const erc20 = ERC20Contract(web3, G$[chainId].address)
-
-//         const allowance = await erc20.methods
-//             .allowance(account, G$ContractAddresses(chainId, 'ExchangeHelper'))
-//             .call()
-//             .then((_: string) => BigNumber.from(_))
-
-//         if (bigInput.lte(allowance)) return
-//         console.log(G$ContractAddresses(chainId, 'ExchangeHelper'))
-
-//         await erc20.methods
-//             .approve(G$ContractAddresses(chainId, 'ExchangeHelper'), MaxUint256.toString())
-//             .send({ from: account })
-//     }
-// }
-
 /**
  * Swap tokens.
  * @param {Web3} web3 Web3 instance.
- * @param {BuyInfo} meta Result of the method getMeta() execution.
+ * @param {SwapInfo} meta Result of the method getMeta() execution.
  * @param {Function} onSent On sent event listener.
  */
 export async function sell(
     web3: Web3,
-    meta: BuyInfo,
+    meta: SwapInfo,
     onSent?: (transactionHash: string, from: string) => void
 ): Promise<TransactionDetails> {
     const chainId = await getChainId(web3)
